making atomic support a multi-cycle operation
The first phase will send out the load request and wait for response. The second phase, on receiving the response will initiate a store request and move the instruction to the next stage.
The atomicity of this will have to be verified. Also the atomic arith is not yet implemented.